# [perf] Cache `Runtime.IsUserType` results